Use class 'divider' to denote columns that logically divide a table
For example, before in the schedule tab there were two separate classes
'matchNumber' and 'lastRedTeam' that had a 1px solid border applied to
on the right to denote a division. Now, that functionality is replaced
by a 'divider' class; however, 'matchNumber' is still a class because
it might prove useful later - it doesn't have any styles right now.