Add colouring for red and blue teams in the schedule tab
I apply a 'redTeam' and 'blueTeam' class to the td's that store a red or
blue team, respectively. Colours for red and blue are then in HSLA
format, which allow the opacity to be changed for alternating rows and
which allow the hue to be changed to create similar shades of red and
blue.