Add colouring for our team and for a team that is searched
Currently, there is no input to actually search a team. There are also
some hacky things going on here, particularly with the nested ternary
operators in defining classes and the usage of !important to override
the more specific "odd row" selector. We might want to look into better
ways to do each of these.